chore: move the shared code for edition state into requestUtils
