+++ frontend/pages/[editionName]/projects.tsx
@@ -14,7 +14,6 @@ import {
   UserRole,
   UUID,
   conflictMapType,
-  Edition,
 } from '../../lib/types';
 import { axiosAuthenticated } from '../../lib/axios';
 import Endpoints from '../../lib/endpoints';
@@ -30,7 +29,7 @@ import FlatList from 'flatlist-react';
 import useUser from '../../hooks/useUser';
 import { SpinnerCircular } from 'spinners-react';
 import Error from '../../components/Error';
-import { getUrlMap, parseError } from '../../lib/requestUtils';
+import {fetchEditionState, getUrlMap, parseError} from '../../lib/requestUtils';
 import RouteProtection from '../../components/RouteProtection';
 import { useRouter } from 'next/router';
 import { NextRouter } from 'next/dist/client/router';
@@ -39,7 +38,6 @@ import useOnScreen from '../../hooks/useOnScreen';
 import PersistLogin from '../../components/PersistLogin';
 import Head from 'next/head';
 import ProjectConflict from '../../components/projects/ProjectConflict';
-import { AxiosInstance, AxiosResponse } from 'axios';
 const magnifying_glass = <FontAwesomeIcon icon={faMagnifyingGlass} />;
 const arrow_out = <Icon icon="bi:arrow-right-circle" />;
 const arrow_in = <Icon icon="bi:arrow-left-circle" />;
@@ -108,77 +106,6 @@ function searchProject(
         setLoading(false);
       }
     });
-}
-
-async function retry_once<T>(
-  func: () => Promise<T>,
-  doSomething: (arg: T) => void,
-  signal: AbortSignal,
-  setError: (error: string) => void,
-  router: NextRouter
-) {
-  try {
-    const result = await func();
-    doSomething(result);
-  } catch (err) {
-    try {
-      const result = await func();
-      doSomething(result);
-    } catch (err) {
-      parseError(err, setError, router, signal);
-    }
-  }
-}
-
-/*
- * Check if there is an active edition, if so then we can compare. If there is no active edition then we know for
- * sure that the current edition is not active.
- */
-export async function load_edition(
-  axiosAuth: AxiosInstance,
-  setEditionActive: (active: boolean) => void,
-  signal: AbortSignal,
-  setError: (error: string) => void,
-  router: NextRouter
-) {
-  await retry_once(
-    async () => {
-      return await axiosAuth.get<Edition>(Endpoints.EDITIONACTIVE);
-    },
-    (response: AxiosResponse<Edition>) => {
-      if (response.data) {
-        const edition = router.query.editionName as string;
-        setEditionActive(edition == response.data.name);
-        return;
-      }
-      setEditionActive(false);
-    },
-    signal,
-    setError,
-    router
-  );
-  return;
-}
-
-export function fetchEditionState(setEditionActive: (v: boolean) => (void), setError: (error: string) => void, router: NextRouter) {
-  useEffect(() => {
-    const controller = new AbortController();
-    const signal = controller.signal;
-    if (router.isReady) {
-      (async () => {
-        await load_edition(
-          axiosAuthenticated,
-          setEditionActive,
-          signal,
-          setError,
-          router
-        );
-      })();
-    }
-    return () => {
-      controller.abort();
-    };
-  }, [router.isReady]);
 }
 
 /**
@@ -311,7 +238,7 @@ const Projects: NextPage = () => {
   const edition = router.query.editionName as string;
 
   let controller = new AbortController();
-  const axiosAuth = useAxiosAuth();
+  useAxiosAuth();
 
   fetchEditionState(setEditionActive, setError, router);
